requirements.txt
pandas
psycopg2-binary
sentence-transformers
numpy 